Migrate financial system to BigDecimal (precision 19, 4) and implement Direct Drop-Ship Loads module

+++ composeApp/src/jvmMain/kotlin/org/safieddine/ablogistics/ui/theme/TitleBar.kt
@@ -161,7 +161,7 @@ fun FrameWindowScope.TitleBar(
                         )
                         val locale = Locale.getDefault()
                         val totalFmt = formatLocalized(
-                            parseLocalizedNumber("%.2f".format(totalFunds), locale), locale
+                            parseLocalizedNumber("%.2f".format(totalFunds.toDouble()), locale), locale
                         )
                         Spacer(Modifier.width(6.dp))
                         Box(
